Add files via upload
The add_service.php file works with a service database table that has the attributes, service_id, service_name, service_price, service_desc, and service_keywords. This is because the service_image attribute was removed temporarily because the image input aspect (which has been commented out in this document) was causing some errors, hence the choice was made and will be rectified once the errors are fixed.